test_requirements.txt
django-nose
coverage
django-coverage
ipdb
flake8
fabric
factory_boy
mock